refactor(host-sdk)!: expose an always-connected messenger to host plugins
Split the symmetric JetWhaleMessenger into a narrow, always-connected base
interface (JetWhaleConnectedMessenger: trySend + request, no send policy) and
the agent-side JetWhaleMessenger that layers OfflineSendPolicy / sendOrQueue /
sendOrFail on top. Host plugins now see only the always-connected face, since a
host plugin instance lives exactly as long as its connection and has nothing to
buffer across. Peer and wire behavior are unchanged; this is an API-surface change.

BREAKING CHANGE: JetWhaleMessagingHostPlugin.messenger and bindMessenger now use
JetWhaleConnectedMessenger instead of JetWhaleMessenger.
