+++ jetwhale-protocol/core/api/jvm/core.api
@@ -129,9 +129,16 @@ public final class com/kitakkun/jetwhale/protocol/messaging/BufferedMessenger : 
 	public final fun close ()V
 	public fun getPayloadFormat ()Lkotlinx/serialization/StringFormat;
 	public fun requestRaw-5_5nbZA (Ljava/lang/String;Ljava/lang/String;Lkotlin/time/Duration;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+	public fun sendRaw (Ljava/lang/String;Ljava/lang/String;)Z
 	public fun sendRaw (Ljava/lang/String;Ljava/lang/String;Lcom/kitakkun/jetwhale/protocol/messaging/OfflineSendPolicy;)Z
 	public final fun startFlush ()V
 	public final fun unbind ()V
+}
+
+public abstract interface class com/kitakkun/jetwhale/protocol/messaging/JetWhaleConnectedMessenger {
+	public abstract fun getPayloadFormat ()Lkotlinx/serialization/StringFormat;
+	public abstract fun requestRaw-5_5nbZA (Ljava/lang/String;Ljava/lang/String;Lkotlin/time/Duration;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+	public abstract fun sendRaw (Ljava/lang/String;Ljava/lang/String;)Z
 }
 
 public final class com/kitakkun/jetwhale/protocol/messaging/JetWhaleConnectionClosedException : com/kitakkun/jetwhale/protocol/messaging/JetWhaleMessagingException {
@@ -157,10 +164,13 @@ public class com/kitakkun/jetwhale/protocol/messaging/JetWhaleMessagingException
 	public synthetic fun <init> (Ljava/lang/String;Ljava/lang/Throwable;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
 }
 
-public abstract interface class com/kitakkun/jetwhale/protocol/messaging/JetWhaleMessenger {
-	public abstract fun getPayloadFormat ()Lkotlinx/serialization/StringFormat;
-	public abstract fun requestRaw-5_5nbZA (Ljava/lang/String;Ljava/lang/String;Lkotlin/time/Duration;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+public abstract interface class com/kitakkun/jetwhale/protocol/messaging/JetWhaleMessenger : com/kitakkun/jetwhale/protocol/messaging/JetWhaleConnectedMessenger {
+	public fun sendRaw (Ljava/lang/String;Ljava/lang/String;)Z
 	public abstract fun sendRaw (Ljava/lang/String;Ljava/lang/String;Lcom/kitakkun/jetwhale/protocol/messaging/OfflineSendPolicy;)Z
+}
+
+public final class com/kitakkun/jetwhale/protocol/messaging/JetWhaleMessenger$DefaultImpls {
+	public static fun sendRaw (Lcom/kitakkun/jetwhale/protocol/messaging/JetWhaleMessenger;Ljava/lang/String;Ljava/lang/String;)Z
 }
 
 public final class com/kitakkun/jetwhale/protocol/messaging/JetWhalePluginPeer {
